+++ Core/src/main/java/net/danh/sinceDungeon/SinceDungeon.java
@@ -84,8 +84,13 @@ public final class SinceDungeon extends JavaPlugin {
         miniMessage = MiniMessage.miniMessage();
         if (configFile == null) configFile = new ConfigManager(this);
         if (languageManager == null) setupLanguage();
-        new ConfigUtils(this, "dungeons/example_dungeon.yml");
-        new ConfigUtils(this, "dungeons/schematic_test_arena.yml");
+
+        boolean regenTemplates = configFile.getConfig().getBoolean("settings.regenerate-default-templates", false);
+        File dungeonsFolder = new File(getDataFolder(), "dungeons");
+        if (regenTemplates || !dungeonsFolder.exists()) {
+            new ConfigUtils(this, "dungeons/example_dungeon.yml");
+            new ConfigUtils(this, "dungeons/schematic_test_arena.yml");
+        }
 
         rewardManager = new RewardManager(this);
         rewardManager.setRewardSystem(new DefaultRewardSystem(this));